third_party/blink/web_tests/external/wpt/editing/run/formatblock_5001-last-expected.txt

This is a testharness.js-based test.
[FAIL] [["formatblock","div"]] "<div><div contenteditable=false><p contenteditable>[foo]</p></div></div>": execCommand("formatblock", false, "div") return value
  assert_equals: expected false but got true
[FAIL] [["formatblock","div"]] "<div><div contenteditable=false><p contenteditable>[foo]</p></div></div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><div contenteditable=\\"false\\"><p contenteditable=\\"\\">foo</p></div></div>" but got "<div><div contenteditable=\\"false\\"><p contenteditable=\\"\\"><div>foo</div></p></div></div>"
[FAIL] [["formatblock","div"]] "<div><div contenteditable=false><p contenteditable>[foo]</p></div></div>" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "" but got "div"
[FAIL] [["formatblock","p"]] "{ <span contenteditable=\\"false\\">A</span> ; <span contenteditable=\\"false\\">B</span> ; <span contenteditable=\\"false\\">C</span> }" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<p><br></p> <span contenteditable=\\"false\\">A</span> ; <span contenteditable=\\"false\\">B</span> ; <span contenteditable=\\"false\\">C</span> " not in array ["<p> <span contenteditable=\\"false\\">A</span> ; <span contenteditable=\\"false\\">B</span> ; <span contenteditable=\\"false\\">C</span> </p>", " <p><span contenteditable=\\"false\\">A</span> ; <span contenteditable=\\"false\\">B</span> ; <span contenteditable=\\"false\\">C</span></p> ", "<p><span contenteditable=\\"false\\">A</span> ; <span contenteditable=\\"false\\">B</span> ; <span contenteditable=\\"false\\">C</span></p>"]
Harness: the test ran to completion.
